+++ src/Models/PgSqlDbConfig.php
@@ -17,41 +17,40 @@ class PgSqlDbConfig extends SqlDbConfig
         return 5432;
     }
 
-    protected function getConnectionFields()
+    public static function getSchema()
     {
-        $fields = parent::getConnectionFields();
-
-        return array_merge($fields, ['charset', 'sslmode', 'timezone', 'application_name']);
-    }
-
-    public static function getDefaultConnectionInfo()
-    {
-        $defaults = parent::getDefaultConnectionInfo();
-        $defaults[] = [
-            'name'        => 'charset',
-            'label'       => 'Character Set',
-            'type'        => 'string',
-            'description' => 'The character set to use for this connection, i.e. ' . static::getDefaultCharset()
-        ];
-        $defaults[] = [
-            'name'        => 'sslmode',
-            'label'       => 'SSL Mode',
-            'type'        => 'string',
-            'description' => 'Enable SSL mode for this connection.'
-        ];
-        $defaults[] = [
-            'name'        => 'timezone',
-            'label'       => 'Timezone',
-            'type'        => 'string',
-            'description' => 'Set the timezone for this connection.'
-        ];
-        $defaults[] = [
-            'name'        => 'application_name',
-            'label'       => 'Application Name',
-            'type'        => 'string',
-            'description' => 'The application used to for monitoring the application with pg_stat_activity.'
+        $schema = parent::getSchema();
+        $extras = [
+            [
+                'name'        => 'charset',
+                'label'       => 'Character Set',
+                'type'        => 'string',
+                'description' => 'The character set to use for this connection, i.e. ' . static::getDefaultCharset()
+            ],
+            [
+                'name'        => 'sslmode',
+                'label'       => 'SSL Mode',
+                'type'        => 'string',
+                'description' => 'Enable SSL mode for this connection.'
+            ],
+            [
+                'name'        => 'timezone',
+                'label'       => 'Timezone',
+                'type'        => 'string',
+                'description' => 'Set the timezone for this connection.'
+            ],
+            [
+                'name'        => 'application_name',
+                'label'       => 'Application Name',
+                'type'        => 'string',
+                'description' => 'The application used to for monitoring the application with pg_stat_activity.'
+            ]
         ];
 
-        return $defaults;
+        $pos = array_search('options', array_keys($schema));
+        $front = array_slice($schema, 0, $pos, true);
+        $end = array_slice($schema, $pos, null, true);
+
+        return array_merge($front, $extras, $end);
     }
 }