+++ src/Models/SqlDbConfig.php
@@ -6,61 +6,48 @@ use DreamFactory\Core\Exceptions\BadRequestException;
 /**
  * SqlDbConfig
  *
- * @property integer $service_id
- * @property array   $connection
- * @property array   $options
- * @property array   $attributes
- * @property array   $statements
- * @method static SqlDbConfig whereServiceId($value)
  */
 class SqlDbConfig extends BaseSqlDbConfig
 {
-    protected $appends = ['host', 'port', 'database', 'username', 'password', 'schema'];
-
     protected $encrypted = ['username', 'password'];
 
     protected $protected = ['password'];
 
-    protected function getConnectionFields()
-    {
-        return ['host', 'port', 'database', 'username', 'password', 'schema'];
-    }
-
-    public static function getDefaultConnectionInfo()
+    public static function getSchema()
     {
         $defaults = [
-            [
+            'host' => [
                 'name'        => 'host',
                 'label'       => 'Host',
                 'type'        => 'string',
                 'description' => 'The name of the database host, i.e. localhost, 192.168.1.1, etc.'
             ],
-            [
+            'port' => [
                 'name'        => 'port',
                 'label'       => 'Port Number',
                 'type'        => 'integer',
                 'description' => 'The number of the database host port, i.e. ' . static::getDefaultPort()
             ],
-            [
+            'database' => [
                 'name'        => 'database',
                 'label'       => 'Database',
                 'type'        => 'string',
                 'description' =>
                     'The name of the database to connect to on the given server. This can be a lookup key.'
             ],
-            [
+            'username' => [
                 'name'        => 'username',
                 'label'       => 'Username',
                 'type'        => 'string',
                 'description' => 'The name of the database user. This can be a lookup key.'
             ],
-            [
+            'password' => [
                 'name'        => 'password',
                 'label'       => 'Password',
                 'type'        => 'password',
                 'description' => 'The password for the database user. This can be a lookup key.'
             ],
-            [
+            'schema' => [
                 'name'        => 'schema',
                 'label'       => 'Schema',
                 'type'        => 'string',
@@ -68,7 +55,7 @@ class SqlDbConfig extends BaseSqlDbConfig
                     'type "default" to only work with the default schema for the given credentials, ' .
                     'or type in a specific schema to use for this service.'
             ],
-//            [
+//            'prefix' => [
 //                'name'        => 'prefix',
 //                'label'       => 'Table Prefix',
 //                'type'        => 'string',
@@ -76,56 +63,15 @@ class SqlDbConfig extends BaseSqlDbConfig
 //            ],
         ];
 
-        return $defaults;
+        return array_merge($defaults, parent::getSchema());
     }
 
-    public static function validateConfig($config, $create = true)
+    public function validate(array $data, $throwException = true)
     {
-        if ($create) {
-            if (empty(array_get($config, 'host')) || empty(array_get($config, 'database'))) {
-                throw new BadRequestException("Database connection information must contain at least host and database name.");
-            }
+        if (empty(array_get($data, 'host')) || empty(array_get($data, 'database'))) {
+            throw new BadRequestException("Database connection information must contain at least host and database name.");
         }
 
-        return parent::validateConfig($config, $create);
-    }
-
-    /**
-     * @param array $schema
-     */
-    protected static function prepareConfigSchemaField(array &$schema)
-    {
-        parent::prepareConfigSchemaField($schema);
-
-        switch ($schema['name']) {
-            case 'options':
-                $schema['label'] = 'Driver Options';
-                $schema['type'] = 'object';
-                $schema['object'] =
-                    [
-                        'key'   => ['label' => 'Name', 'type' => 'string'],
-                        'value' => ['label' => 'Value', 'type' => 'string']
-                    ];
-                $schema['description'] = 'A key-value array of driver-specific connection options.';
-                break;
-            case 'attributes':
-                $schema['label'] = 'Driver Attributes';
-                $schema['type'] = 'object';
-                $schema['object'] =
-                    [
-                        'key'   => ['label' => 'Name', 'type' => 'string'],
-                        'value' => ['label' => 'Value', 'type' => 'string']
-                    ];
-                $schema['description'] =
-                    'A key-value array of attributes to be set after connection.' .
-                    ' For further information, see http://php.net/manual/en/pdo.setattribute.php';
-                break;
-            case 'statements':
-                $schema['label'] = 'Additional SQL Statements';
-                $schema['type'] = 'array';
-                $schema['items'] = 'string';
-                $schema['description'] = 'An array of SQL statements to run during connection initialization.';
-                break;
-        }
+        return parent::validate($data);
     }
 }